feat(xposed): add libxposed entry point alongside de.robv
Prototype of dual-API support. The de.robv API has been unmaintained since
2016 (de.robv.android.xposed:api stops at 82, published 2016-04-15), while
io.github.libxposed:api is the API modern frameworks implement natively --
current frameworks service the old one through a legacy shim.

Rather than migrating, ship both entry points:

  assets/xposed_init                  -> AnyWebView        (de.robv)
  META-INF/xposed/java_init.list      -> AnyWebViewModule  (libxposed)

A straight migration is not viable here: the libxposed 101 AAR declares
minSdkVersion 26, so adopting it alone would drop API 24-25, and every
framework that does not implement libxposed would lose the module entirely.
Both dependencies are compileOnly, which keeps the libxposed AAR off the
runtime graph and therefore keeps its minSdkVersion 26 out of the manifest
merge; the packaged APK still reports minSdkVersion 24.

All logic moves to WebViewProviderInjector, which has no Xposed dependency of
any kind. The two adapters only locate SystemImpl.getWebViewPackages(), install
a hook, and hand the result to the injector. AnyWebView drops from 258 lines to
48.

The libxposed adapter is the simpler of the two: onSystemServerStarting() fires
only for system_server, so no package-name filtering is needed, and the
interceptor model expresses the append directly -- proceed(), then return a
longer array -- instead of mutating a result out of band.

Double-hooking is prevented by construction rather than by framework behaviour:
if a framework loads both entry points, each injector dedupes candidates
against the WebViewProviderInfo[] it is handed, so whichever hook runs second
sees the first one's additions already present and contributes nothing.

Prototype status: builds, lints, and the packaged APK carries both entry points
with minSdkVersion 24 and no bundled libxposed classes. Not yet run on a
device, under either API.
