+++ app/src/main/resources/META-INF/xposed/java_init.list
@@ -1,0 +1,1 @@
+com.thinkdifferent.anywebview.AnyWebViewModule
